+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] / 2026-01-23
+### Features
+- Support download bundle with parameters in the url and local file path.
+- Support download from redirect url and validate zip file. `BundleUri`
+### Fixes
+- Fix `Console` not supporting colors in some terminals. (Fix: #7)
+### Updates
+- Update `NameAndVersionBundleUtils` with `RemoveVersionBundle` to remove version from bundle file name.
+- Update `BundleDownloadUtils` to support download url with parameters and local file path.
+- Update `BundleDownloadUtils` to use `BundleUri` with handling local file or url.
+- Update `BundleUri` to validate `zip` file when downloading from url.
+### Tests
+- Add `BundleDownloadTests` to test download url with parameters.
+- Add `UriTests` to test uri with local files or urls.
+
 ## [1.0.5] / 2025-11-03
 ### Features
 - Support install `{name}.{version}.bundle.zip` format and remove `{version}` when install bundle.
@@ -61,6 +76,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `DownloadUtils` to support local file.
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.1.0]: ../../compare/1.0.5...1.1.0
 [1.0.5]: ../../compare/1.0.4...1.0.5
 [1.0.4]: ../../compare/1.0.3...1.0.4
 [1.0.3]: ../../compare/1.0.2...1.0.3